--- src/Form/fields/ObjectField/ObjectField.css
@@ -1,8 +1,0 @@
-.toggleButton {
-  width: 100%;
-  margin-bottom: 0.625rem;
-}
-
-:global(.card-container > .ant-tabs-card > .ant-tabs-content) > .tab {
-  padding-top: 0.625rem;
-}
