--- src/Form/fields/ObjectField/ObjectField.css.d.ts
@@ -1,2 +1,0 @@
-export const toggleButton: string;
-export const tab: string;
