+++ src/Form/fields/ObjectField/components/Fieldset.css
@@ -3,3 +3,12 @@
   padding: 0.625rem 0.625rem 0;
   margin-bottom: 0.625rem;
 }
+
+.toggleButton {
+  width: 100%;
+  margin-bottom: 0.625rem;
+}
+
+:global(.card-container > .ant-tabs-card > .ant-tabs-content) > .tab {
+  padding-top: 0.625rem;
+}
