+++ src/Form/fields/ObjectField/components/Fieldset.css.d.ts
@@ -1,1 +1,3 @@
 export const fieldset: string;
+export const toggleButton: string;
+export const tab: string;
